requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
sqlparse==0.5.3
